+++ click_extended/decorators/__init__.py
@@ -1,5 +1,7 @@
 """Initialization file for the `click_extended.decorators` module."""
 
+from click_extended.decorators.at_least import at_least
+from click_extended.decorators.at_most import at_most
 from click_extended.decorators.exclusive import exclusive
 from click_extended.decorators.random_bool import random_bool
 from click_extended.decorators.random_choice import random_choice
@@ -26,6 +28,8 @@ from click_extended.decorators.to_case import (
 )
 
 __all__ = [
+    "at_least",
+    "at_most",
     "exclusive",
     "random_bool",
     "random_choice",